requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.10.0
playwright==1.49.0
requests==2.32.3
lxml==6.0.2
xlrd==2.0.2
pandas==2.2.3
